+++ lib/routes/taoguba/blog.ts
@@ -1,9 +1,7 @@
 import { Route } from '@/types';
-import cache from '@/utils/cache';
 import got from '@/utils/got';
 import { load } from 'cheerio';
-import timezone from '@/utils/timezone';
-import { parseDate } from '@/utils/parse-date';
+import { rootUrl, renderPostDetail } from './util';
 
 export const route: Route = {
     path: ['/blog/:id', '/user/:id'],
@@ -20,7 +18,7 @@ export const route: Route = {
     },
     radar: [
         {
-            source: ['taoguba.com.cn/blog/:id', 'taoguba.com.cn/'],
+            source: ['tgb.cn/blog/:id', 'tgb.cn/'],
         },
     ],
     name: '用户博客',
@@ -31,7 +29,6 @@ export const route: Route = {
 async function handler(ctx) {
     const id = ctx.req.param('id');
 
-    const rootUrl = 'https://www.taoguba.com.cn';
     const currentUrl = `${rootUrl}/blog/${id}`;
 
     const response = await got({
@@ -51,53 +48,13 @@ async function handler(ctx) {
             const a = item.find('a').first();
 
             return {
-                title: a.text(),
+                title: a.text().trim(),
                 link: `${rootUrl}/${a.attr('href')}`,
                 author,
             };
         });
 
-    items = await Promise.all(
-        items.map((item) =>
-            cache.tryGet(item.link, async () => {
-                const detailResponse = await got({
-                    method: 'get',
-                    url: item.link,
-                });
-                if (detailResponse.url?.startsWith('https://www.taoguba.com.cn/topic/transfer')) {
-                    item.description = '登录后查看完整文章';
-                    return item;
-                }
-
-                const content = load(detailResponse.data);
-
-                content('#videoImg').remove();
-                content('img').each((_, img) => {
-                    if (img.attribs.src2) {
-                        img.attribs.src = img.attribs.src2;
-                        delete img.attribs.src2;
-                        delete img.attribs['data-original'];
-                    }
-                });
-
-                item.description = content('#first').html();
-                item.pubDate = timezone(
-                    parseDate(
-                        content('.article-data span')
-                            .eq(1)
-                            .text()
-                            .match(/\d{4}-\d{2}-\d{2} \d{2}:\d{2}/)
-                    ),
-                    +8
-                );
-                item.category = content('.article-topic-list span')
-                    .toArray()
-                    .map((item) => $(item).text().trim());
-
-                return item;
-            })
-        )
-    );
+    items = await Promise.all(items.map(async (item) => await renderPostDetail(item)));
 
     return {
         title: `淘股吧 - ${author}`,
