+++ lib/routes/taoguba/index.ts
@@ -1,13 +1,11 @@
 import { Route } from '@/types';
-import cache from '@/utils/cache';
 import got from '@/utils/got';
 import { load } from 'cheerio';
-import timezone from '@/utils/timezone';
-import { parseDate } from '@/utils/parse-date';
+import { rootUrl, renderPostDetail } from './util';
 
 export const route: Route = {
     path: ['/index', '/:category?'],
-    name: 'Unknown',
+    name: '论坛帖子',
     maintainers: ['nczitzk'],
     handler,
     description: `| 淘股论坛 | 社区总版 | 精华加油 | 网友点赞 |
@@ -18,7 +16,6 @@ export const route: Route = {
 async function handler(ctx) {
     const category = ctx.req.param('category') ?? 'zongban';
 
-    const rootUrl = 'https://www.taoguba.com.cn';
     const currentUrl = `${rootUrl}/${category}/`;
 
     const response = await got({
@@ -28,63 +25,22 @@ async function handler(ctx) {
 
     const $ = load(response.data);
 
-    let items = $('.items-comment-list')
+    let items = $('.Nbbs-tiezi-lists')
         .slice(0, ctx.req.query('limit') ? Number.parseInt(ctx.req.query('limit')) : 70)
         .toArray()
         .map((item) => {
             item = $(item);
 
-            const a = item.find('.items-list-tittle a');
-            a.find('b').remove();
+            const a = item.find('.middle-list-tittle a');
 
             return {
-                title: a.text(),
+                title: a.text().trim(),
                 link: `${rootUrl}/${a.attr('href')}`,
-                author: item.find('.items-list-user a').text().trim(),
+                author: item.find('.middle-list-user a').text().trim(),
             };
         });
 
-    items = await Promise.all(
-        items.map((item) =>
-            cache.tryGet(item.link, async () => {
-                const detailResponse = await got({
-                    method: 'get',
-                    url: item.link,
-                });
-                if (detailResponse.url?.startsWith('https://www.taoguba.com.cn/topic/transfer')) {
-                    item.description = '登录后查看完整文章';
-                    return item;
-                }
-
-                const content = load(detailResponse.data);
-
-                content('#videoImg').remove();
-                content('img').each((_, img) => {
-                    if (img.attribs.src2) {
-                        img.attribs.src = img.attribs.src2;
-                        delete img.attribs.src2;
-                        delete img.attribs['data-original'];
-                    }
-                });
-
-                item.description = content('#first').html();
-                item.pubDate = timezone(
-                    parseDate(
-                        content('.article-data span')
-                            .eq(1)
-                            .text()
-                            .match(/\d{4}-\d{2}-\d{2} \d{2}:\d{2}/)
-                    ),
-                    +8
-                );
-                item.category = content('.article-topic-list span')
-                    .toArray()
-                    .map((item) => $(item).text().trim());
-
-                return item;
-            })
-        )
-    );
+    items = await Promise.all(items.map(async (item) => await renderPostDetail(item)));
 
     return {
         title: $('head title').text().trim().split('_')[0],
