+++ lib/routes/taoguba/namespace.ts
@@ -2,6 +2,6 @@ import type { Namespace } from '@/types';
 
 export const namespace: Namespace = {
     name: '淘股吧',
-    url: 'taoguba.com.cn',
+    url: 'tgb.cn',
     lang: 'zh-CN',
 };
